ci: enforce format:check and fail on lint warnings
CI ran typecheck/lint/test/build but never `format:check`, and `lint` had no
`--max-warnings` cap, so unformatted code and lint warnings passed. Adds a
Format check step and moves `--max-warnings 0` into the `lint` script itself
so local and CI agree.

Clears the 13-warning baseline first, since the gate would fail on it:

- memory-reader / theme-init: unused `catch (e)` bindings → bare `catch`
- ActivityChart: `buildBuckets` has no other consumer → unexported
- CreateFormKit: split into `CreateFormKit.tsx` (components only) and
  `formKit.ts` (tool catalog, limits, `openDocs`/`validateName`/
  `useCreateFormKeys`) — react-refresh/only-export-components
- useTheme: same split, `useTheme.ts` (types + context + hook) and
  `ThemeProvider.tsx` (the provider component)
- ProjectOverview: the ⌘F/⌘P key handler was missing `openGlobalSearch`;
  the search openers are now `useCallback` so the dep is stable

Refs #168
